Adopt type states to manage bus sharing
In https://github.com/japaric/embedded-hal/issues/35 one of the
proposals for dealing with sharing the buses is to use this pattern.

The most appealing aspect of this is that it doesn't require any
specific knowledge of synchronization or concurrency used by
the embedding application.

The `Owned::borrow` method feels a little warty and un-ergonomic.
Ideally we'd be able to use `Deref` and `DerefMut` to have the
compiler automatically `borrow()` when doing something like
`owned.get_bank_a_data()`, but it doesn't appear to be possible
to have it meet our lifetime requirements.